# specs/002-enterprise-uiux-overhaul/contracts/gateway-uiux.openapi.yaml
openapi: 3.0.3
info:
  title: Orbit Tasks Gateway v2 Contract
  version: 2.0.0
  description: Public REST contract for Sprint-DSU loop, work item multiview CRUD, collaboration triage, and schedule intelligence.
servers:
  - url: https://tasksapi.infinitefallcult.trade
    description: Production gateway
  - url: http://localhost:8080
    description: Local gateway
paths:
  /api/v2/work-items:
    get:
      summary: List work items
      operationId: listWorkItemsV2
      parameters:
        - in: query
          name: projectId
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: assignee
          schema:
            type: string
        - in: query
          name: query
          schema:
            type: string
      responses:
        "200":
          description: Work item list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkItem"
    post:
      summary: Create work item
      operationId: createWorkItemV2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkItemRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItem"
  /api/v2/work-items/dependency-graph:
    get:
      summary: Fetch dependency graph
      operationId: getDependencyGraphV2
      parameters:
        - in: query
          name: projectId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dependency graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DependencyGraph"
  /api/v2/work-items/{workItemId}:
    patch:
      summary: Patch work item
      operationId: patchWorkItemV2
      parameters:
        - $ref: "#/components/parameters/WorkItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchWorkItemRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItem"
  /api/v2/work-items/{workItemId}/status:
    patch:
      summary: Patch status only
      operationId: patchWorkItemStatusV2
      parameters:
        - $ref: "#/components/parameters/WorkItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItem"
  /api/v2/work-items/{workItemId}/dependencies:
    post:
      summary: Add dependency
      operationId: addDependencyV2
      parameters:
        - $ref: "#/components/parameters/WorkItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [toWorkItemId]
              properties:
                toWorkItemId:
                  type: string
                  format: uuid
                type:
                  type: string
      responses:
        "200":
          description: Created dependency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DependencyEdge"
  /api/v2/dependencies/{dependencyId}:
    delete:
      summary: Delete dependency
      operationId: deleteDependencyV2
      parameters:
        - in: path
          name: dependencyId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
  /api/v2/work-items/{workItemId}/activity:
    get:
      summary: List work item activity
      operationId: listWorkItemActivityV2
      parameters:
        - $ref: "#/components/parameters/WorkItemId"
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkItemActivity"
...
components:
  parameters:
    WorkItemId:
      in: path
      name: workItemId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    WorkItem:
      type: object
      required: [workItemId, projectId, type, title, status, createdAt, updatedAt]
      properties:
        workItemId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        status:
          type: string
        assignee:
          type: string
          nullable: true
        startAt:
          type: string
          nullable: true
        dueAt:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
        estimateMinutes:
          type: integer
          nullable: true
        actualMinutes:
          type: integer
          nullable: true
        blockedReason:
          type: string
          nullable: true
        markdownBody:
          type: string
          nullable: true
        activityCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    CreateWorkItemRequest:
      type: object
      required: [projectId, type, title]
      properties:
        projectId:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        assignee:
          type: string
        startAt:
          type: string
        dueAt:
          type: string
        priority:
          type: string
        estimateMinutes:
          type: integer
        actualMinutes:
          type: integer
        blockedReason:
          type: string
        markdownBody:
          type: string
    PatchWorkItemRequest:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: string
        assignee:
          type: string
          nullable: true
        startAt:
          type: string
          nullable: true
        dueAt:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
        estimateMinutes:
          type: integer
          nullable: true
        actualMinutes:
          type: integer
          nullable: true
        blockedReason:
          type: string
          nullable: true
        markdownBody:
          type: string
          nullable: true
    DependencyEdge:
      type: object
      properties:
        dependencyId:
          type: string
        fromWorkItemId:
          type: string
        toWorkItemId:
          type: string
        type:
          type: string
        createdAt:
          type: string
    DependencyGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
            properties:
              workItemId:
                type: string
              title:
                type: string
              status:
                type: string
              upstreamCount:
                type: integer
              downstreamCount:
                type: integer
        edges:
          type: array
          items:
            $ref: "#/components/schemas/DependencyEdge"
    WorkItemActivity:
      type: object
      properties:
        activityId:
          type: string
        workItemId:
          type: string
        action:
          type: string
        actorId:
          type: string
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
...